Replace Desktop.getDesktop().browse()
As a result of bad Linux support I decided to rewrite this to not use
the Desktop class.

Implemented a custom class "Browser" that executes a system call to open
a browser. It's supposed to work on Windows, MacOS and Linux, but is yet
to be tested under anything other than Linux.