Fix searching for placeholder text when empty
When there's no search text a placeholder is added through the custom
behavior of ExtTextField. This had a bug causing the `getText` method to
return the placeholder text when it should return empty string.

Fixed by keeping track of placeholder status and overriding `getText` to
return empty string in case the placeholder is active.